pyproject.toml
[tool.poetry]
name = "satellite_tracker"
version = "0.1.0"
description = ""
authors = ["Matt Seminara <matthewseminara@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.4.2"
requests = "^2.27.1"
ipykernel = "^6.13.0"
boto3 = "^1.24.1"
apache-airflow = "^2.3.2"
apache-airflow-providers-amazon = "^3.4.0"
apache-airflow-providers-postgres = "^4.1.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
